Jsm add interactive mode

+++ parcel/segment.py
@@ -2,8 +2,11 @@ from intervaltree import Interval, IntervalTree
 import os
 import tempfile
 import pickle
+from portability import OS_WINDOWS
+import string
+import random
 
-if os.name == 'nt':
+if OS_WINDOWS:
     WINDOWS = True
     from Queue import Queue
 else:
@@ -16,6 +19,7 @@ else:
 from log import get_logger
 from utils import get_pbar, md5sum, mmap_open, set_file_length,\
     get_file_type, STRIP
+from const import SAVE_INTERVAL
 from progressbar import ProgressBar, Percentage, Bar, ETA
 
 log = get_logger('segment')
@@ -24,7 +28,7 @@ log = get_logger('segment')
 class SegmentProducer(object):
 
     def __init__(self, file_id, file_path, save_path, load_path, n_procs, size,
-                 save_interval=int(1e6), check_segment_md5sums=False):
+                 save_interval=SAVE_INTERVAL, check_segment_md5sums=False):
 
         self.file_id = file_id
         self.file_path = file_path
@@ -75,9 +79,10 @@ class SegmentProducer(object):
     def validate_segment_md5sums(self):
         if not self.check_segment_md5sums:
             return True
+        corrupt_segments = 0
         intervals = sorted(self.completed.items())
         pbar = ProgressBar(widgets=[
-            'Checksumming {}:'.format(self.file_id), Percentage(), ' ',
+            'Checksumming {}: '.format(self.file_id), Percentage(), ' ',
             Bar(marker='#', left='[', right=']'), ' ', ETA()])
         with mmap_open(self.file_path) as data:
             for interval in pbar(intervals):
@@ -91,9 +96,13 @@ class SegmentProducer(object):
                 chunk = data[interval.begin:interval.end]
                 checksum = md5sum(chunk)
                 if checksum != interval.data.get('md5sum'):
-                    log.warn('Redownloading corrupt segment {}, {}.'.format(
+                    log.debug('Redownloading corrupt segment {}, {}.'.format(
                         interval, checksum))
+                    corrupt_segments += 1
                     self.completed.remove(interval)
+        if corrupt_segments:
+            log.warn('Redownloading {} currupt segments.'.format(
+                corrupt_segments))
 
     def load_state(self, load_path, size):
         # Establish default intervals
@@ -153,8 +162,35 @@ class SegmentProducer(object):
             temp.flush()
             os.fsync(temp.fileno())
             temp.close()
-            # Rename temp file as our save file
-            os.rename(temp.name, self.save_path)
+
+            # Rename temp file as our save file, this could fail if
+            # the state file and the temp directory are on different devices
+            if OS_WINDOWS and os.path.exists(self.save_path):
+                # If we're on windows, there's not much we can do here
+                # except stash the old state file, rename the new one,
+                # and back up if there is a problem.
+                old_path = os.path.join(tempfile.gettempdir(), ''.join(
+                    random.choice(string.ascii_lowercase + string.digits)
+                    for _ in range(10)))
+                try:
+                    # stash the old state file
+                    os.rename(self.save_path, old_path)
+                    # move the new state file into place
+                    os.rename(temp.name, self.save_path)
+                    # if no exception, then delete the old stash
+                    os.remove(old_path)
+                except Exception as msg:
+                    log.error('Unable to write state file: {}'.format(msg))
+                    try:
+                        os.rename(old_path, self.save_path)
+                    except:
+                        pass
+                    raise
+            else:
+                # If we're not on windows, then we'll just try to
+                # atomically rename the file
+                os.rename(temp.name, self.save_path)
+
         except KeyboardInterrupt:
             log.warn('Keyboard interrupt. removing temp save file'.format(
                 temp.name))
